:truck: move html to own folder
