+++ html/floorball.html
@@ -72,12 +72,12 @@
   <div id="details-modal"></div>
   <div id="delete-all-modal"></div>
   <hr />
-  <script type="module" src="./index.js"></script>
+  <script type="module" src="./setup.js"></script>
   <script type="module">
     import {
-      index
-    } from "./index.js";
-    index("floorball");
+      setup
+    } from "./setup.js";
+    setup("floorball");
   </script>
   <!-- bootstrap -->
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/js/bootstrap.bundle.min.js" integrity="sha384-JEW9xMcG8R+pH31jmWH6WWP0WintQrMb4s7ZOdauHnUtxwoG2vI5DkLtS3qm9Ekf" crossorigin="anonymous"></script>
