+++ html/hockey.html
@@ -43,7 +43,7 @@
 
   <!-- Bootstrap Icons -->
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.4.0/font/bootstrap-icons.css" />
-  <link rel="stylesheet" href="./index.css" />
+  <link rel="stylesheet" href="/index.css" />
 </head>
 
 <body>
@@ -72,12 +72,12 @@
   <div id="details-modal"></div>
   <div id="delete-all-modal"></div>
   <hr />
-  <script type="module" src="./index.js"></script>
+  <script type="module" src="/setup.js"></script>
   <script type="module">
     import {
-      index
-    } from "./index.js";
-    index("hockey");
+      setup
+    } from "/setup.js";
+    setup("hockey");
   </script>
   <!-- bootstrap -->
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/js/bootstrap.bundle.min.js" integrity="sha384-JEW9xMcG8R+pH31jmWH6WWP0WintQrMb4s7ZOdauHnUtxwoG2vI5DkLtS3qm9Ekf" crossorigin="anonymous"></script>
